Refactor and correct VacanteController & Livewire components
This commit changes the naming convention of 'vacanteController' to the appropriate standard 'VacanteController'. Also refactored editVacante blade file to use new 'imagen_nueva' as a model instead of 'imagen', for clear differentiation between new and old images.

In Livewire component, editVacante, several changes were made. The most significant was introducing the 'imagen_nueva' property that holds the new image. Also, mount function is updated to include 'vacante_id' property for it to be later used in 'editarVacante' function. This function updated to now handle the scenario when the image is updated. Also, the original 'imagen' property is updated based on the presence of 'imagen_nueva'.

Also, in web.php file the route names were corrected from 'vacanteController' to the standardised 'VacanteController'. This is a necessary update to keep coding standards and conventions.

In VacanteController, the unnecessary functions were removed and replaced with an edit function including authorisation. Also updated with the corrected class name.

In app/Providers/AuthServiceProvider.php, the policy for Vacante was added to the 'policies' property in an effort to standardise access control across the application.

In CrearVacante component, comments were added for clarity. Function 'crearVacante' was simplified for easier understanding and better flow.

Overall, these adjustments aim to improve code readability and maintainability.