Create new show.blade, update error messages and modifies dependencies in composer.lock
This commit adds a new show.blade.php file that displays detailed information of a specific 'vacante'. The error regarding the requirement of salario has also been corrected to reflect that only 'salario_id' is necessary instead of 'the salario is required'. Furthermore, changes are made in composer.lock file to modify project dependencies, bringing in multiple packages for different functionalities to support the project's evolving needs.