"Add job application feature in Livewire component"
This commit introduces a new feature allowing users to apply for a job vacancy directly from the website. A new Livewire component `PostularVacante` has been created to handle applications, it includes a file upload feature for CVs in PDF format, and methods to validate, store, and associate the CV with a job vacancy and the user applying. A success message is returned to the user upon successful submission. Changes also include necessary modifications to the `Vacante` Model to define its relationship with `Candidato` Model. Views were also updated correspondingly to handle new functionality. This new feature enhances the user experience by providing a direct and streamlined application process.