Refactored notification and vacancy model functions
In "app/Notifications/NuevoCandidato.php", the toArray method has been refactored to a toDatabase method for clearer function naming and to properly indicate that this function is meant to save notifications to the database.

In the "app/Models/Vacante.php", apart from defining what fields are fillable and cast type for specific attributes, detailed methods descriptions and relations with other models were added to increase clarity and ease-of-understanding for people working with this model. Other developers should now immediately understand these model's methods' purpose and functionality.