Add notification feature for new job applicant
New properties were introduced in `NuevoCandidato.php` construct. The `via` method was updated to include `database` besides `mail`. Changes were made in `toMail` and `toDatabase` methods to handle the new attributes. `PostularVacante.php` was updated to notify the recruiter when a new candidate applied for a job. Route to handle notifications was added in `web.php`. An index blade file was created to display notifications. Finally, the return method of `candidatos` in `Vacante.php` was fixed. These changes allow a recruiter to be notified when a new candidate applies for a job.